requirements.txt
Django==5.1.1
Pillow==10.4.0
python-decouple==3.8
coverage==7.6.3